// reEnergyUI/src/app/pages/model/center/CenGereralForm.model.ts
export class CenterGeneralinfoSection {
  basicCertification = new CenterBasicCertification();
  generalInfo = new CenterGeneralInfo();
  organizationInformation = new CenterOrganizationInformation();
}


export class CenterBasicCertification {
  organization_type: string;
  ayush_system: string;
}

export class CenterOrganizationhead {
  org_head_same_as_spoc: boolean;
  spoc_name: string;
  salutation: string;
  spoc_mob: string;
  spoc_desig: string;
  spoc_email: string;
}
// export class CenterCertificatoncordinatr {
//   coord_name: string;
//   coord_mob: string;
//   coord_desig: string;
//   coord_email: string;
// }

export class CenterSplitLocation {
  id: number;
  split_build_name: string;
  split_state: string;
  split_district: number;
  split_city: string;
  split_address: string;
  split_pincode: string;
  name_spoc_loc: string;
  mob_spoc_loc: string;
  email_spoc_loc: string;
  dist_frm_main_loc: string;
  district_list:CentralDistrict=new CentralDistrict()
}
class CentralDistrict {
  id: number;
  districtname: string;
  statecode: string;
  gstcode: string;
}

export class CenterLocation {
  latitude: string;
  longitude: string;
  location_of_organization: string;
  does_org_hv_split_loc: boolean;
  split_loc_nos: number;
  splitLocation = new Array<CenterSplitLocation>();
}

export class CenterGeneralInfo {
  org_name: string;
  pat_stay_overnight: boolean;
  build_name: string;
  address: string;
  street: string;
  state: string;
  district: string;
  city: string;
  pincode: string;
  nearby_landmark: string;
  org_website: string;
  total_sanctioned_beds: string;
  total_operational_beds: string;
  no_beds: string;
  landline: string;
  spoc_name: string;
  spoc_mob: string;
  spoc_designtion: string;
  spoc_email: string;
  organizationhead = new CenterOrganizationhead();
  // certificatoncordinatr = new CenterCertificatoncordinatr();
  location = new CenterLocation();
}

export class CenterOrganizationIsAlreadyEmpanelledWith {
  cghs: boolean;
  cghs_url: string;
  railways: boolean;
  railways_url: string;
  ayushmanbharat: boolean;
  ayushmanbharat_url: string;
  publichealthinsuranceschemes: boolean;
  publichealthinsuranceschemes_url: string;
  echs: boolean;
  echs_url: string;
  stategovernmenthealthscheme: boolean;
  stategovernmenthealthscheme_url: string;
  others: boolean;
  others_url: string;
  other_scheme_organization_is_linked: string;
  mentionothertypeofschemes_url: string;
  none: boolean;
}

export class CenterOrganizationInformation {
  establishment_date_of_org: string;
  organization_ownership_type: string;
  id: number;
  type: string;
  mention_ownership_type: string;
  mention_other_scheme_orgztn_linked: string;
  month_year_setup: string;
  organization_ownership_type_url: string;
  organization_is_already_empanelled_with = new CenterOrganizationIsAlreadyEmpanelledWith();
}